First GUI created for the Navigation of the drone
This file is a tkinter GUI used to visualize the path algorithm of the drone. To use, click on the "outline" of the field or farm that the drone will scan and the horizontal paths will appear in the window.
NOTE: This GUI has a lot of dead code in it because it was modified from a previous project.